fix(web): keep forum feed model coverable
Keep the shared feed types in the executable transition module so the changed-file coverage contract can observe the source boundary.
